LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-axis-writing-mode-expected.txt


PASS Initial axis
FAIL Vertical axis assert_equals: expected "175px" but got "100px"
PASS Horizontal axis
PASS Block axis in horizontal writing-mode
PASS Inline axis in horizontal writing-mode
FAIL Block axis in vertical writing-mode assert_equals: expected "125px" but got "100px"
FAIL Inline axis in vertical writing-mode assert_equals: expected "175px" but got "100px"
